Removed adapters for Grid and List RecyclerView and use one general RecyclerView adapter instead (using CardViewMovieAdapter as a base and renamed it as MovieAdapter). Also renamed ListMovieHolder to MovieViewHolder.
Implemented SortedList and its Callbacks in the MovieAdapter in order to neatly implement the SearchView feature. Now it is possible to search movies by title, year, release date, director, or stars.

Modified the sortBy method to compensate the change in the adapter and the use of SortedList. Simplified and generalized the showRecycler... method in MainActivity to reduce redundant codes.

Added ripple effect to the spinner. Removed some unused & unnecessary drawables.

Flattened a lot of layout, especially the ones that use ConstraintLayout.
Movie poster in the MovieDetailsActivity is now contained in a CardView.